djangoServer/requirements.txt
Django==2.1.2
djangorestframework==3.9.0
numpy==1.15.3
Pillow==5.3.0
scipy==1.1.0
Keras==2.2.4
tensorflow==1.11.0
django-cors-headers==2.4.0
